Revert "Camilla stringss"
