+++ app/src/main/res/layout/activity_chat.xml
@@ -97,7 +97,7 @@
             android:id="@+id/messageInput"
             android:layout_width="300dp"
             android:layout_height="wrap_content"
-            android:hint="@string/TypeMessage"
+            android:hint="Type message here"
             android:inputType="text|textLongMessage|textAutoComplete|textAutoCorrect|textImeMultiLine"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
@@ -109,7 +109,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:backgroundTint="@color/black"
-            android:text="@string/Send"
+            android:text="Send"
             app:cornerRadius="10dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
