+++ build.gradle.kts
@@ -2,8 +2,7 @@
 plugins {
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.kotlin.android) apply false
-   alias(libs.plugins.google.gms.google.services) apply false
-
+    alias(libs.plugins.google.gms.google.services) apply false
 }
